move csv diff tool from cli/utils to libs/qa. add a bunch of comments.
