+++ cli/utils.py
@@ -2,19 +2,18 @@ import logging
 import pathlib
 import subprocess
 from datetime import datetime
-from typing import Optional
 
 import click
 import structlog
 
 from covidactnow.datapublic import common_df
-from covidactnow.datapublic.common_fields import CommonFields, COMMON_FIELDS_TIMESERIES_KEYS
+from covidactnow.datapublic.common_fields import CommonFields
 from libs import github_utils
 from libs.datasets import combined_datasets
-from pydantic import BaseModel
 import pandas as pd
 import numpy as np
 
+from libs.qa.common_df_diff import DatasetDiff
 
 _logger = logging.getLogger(__name__)
 
@@ -90,113 +89,6 @@ def form_path_name(csv_path_format, output_dir):
         timestamp=datetime.now().strftime("%Y%m%dT%H%M%S"),
     )
     return csv_path
-
-
-class DatasetDiff(BaseModel):
-    duplicates_dropped: pd.DataFrame
-    melt: pd.DataFrame
-    all_variable_fips: pd.MultiIndex
-    my_ts: Optional[pd.MultiIndex] = None
-    common_fips: Optional[pd.DataFrame] = None
-    my_ts_points: Optional[pd.DataFrame] = None
-    ts_diffs: Optional[pd.DataFrame] = None
-
-    class Config:
-        arbitrary_types_allowed = True
-
-    def __str__(self):
-        return f"""Duplicate rows in this file: {self.duplicates_dropped}
-TS only in this file: {self.my_ts}
-TS points only in this file: {self.my_ts_points.groupby('date').size().to_dict()}
-TS diffs: {self.ts_diffs if self.ts_diffs is not None else ''}
-TS diffs: {self.ts_diffs.groupby('variable has_overlap'.split()).mean() if self.ts_diffs is not None else ''}
-"""
-
-    @staticmethod
-    def make(df: pd.DataFrame) -> "DatasetDiff":
-        dups = df.loc[df.index.duplicated(keep=False)]
-        if not dups.empty:
-            df = df.drop_duplicates()
-
-        df = df.reset_index()
-        columns_to_drop = {
-            "index",
-            CommonFields.STATE,
-            CommonFields.COUNTRY,
-            CommonFields.AGGREGATE_LEVEL,
-        }.intersection(df.columns)
-        for col in df.select_dtypes(include="object"):
-            if col != "fips":
-                print(f"dropping based on type {col}")
-                columns_to_drop.add(col)
-        if columns_to_drop:
-            df = df.drop(columns=columns_to_drop)
-
-        melt = (
-            df.melt(id_vars=COMMON_FIELDS_TIMESERIES_KEYS)
-            .set_index(["variable"] + COMMON_FIELDS_TIMESERIES_KEYS)
-            .dropna()
-        )
-
-        all_variable_fips = melt.groupby("variable fips".split()).first().index
-        return DatasetDiff(duplicates_dropped=dups, melt=melt, all_variable_fips=all_variable_fips)
-
-    def compare(self, other: "DatasetDiff"):
-        # Index of <variable, fips> that have at least one real value in only one dataset
-        self.my_ts = self.all_variable_fips.difference(other.all_variable_fips)
-        other.my_ts = other.all_variable_fips.difference(self.all_variable_fips)
-
-        common_variable_fips = self.all_variable_fips.intersection(other.all_variable_fips)
-        self.common_fips = self.melt.loc[
-            self.melt.reset_index(CommonFields.DATE).index.isin(common_variable_fips)
-        ]
-        other.common_fips = other.melt.loc[
-            other.melt.reset_index(CommonFields.DATE).index.isin(common_variable_fips)
-        ]
-
-        joined_ts = pd.merge(
-            self.common_fips["value"],
-            other.common_fips["value"],
-            how="outer",
-            left_index=True,
-            right_index=True,
-            suffixes=("_l", "_r"),
-        )
-        joined_ts_notna = joined_ts.notna()
-        self.my_ts_points = joined_ts.loc[
-            joined_ts_notna["value_l"] & ~joined_ts_notna["value_r"], "value_r"
-        ]
-        other.my_ts_points = joined_ts.loc[
-            joined_ts_notna["value_r"] & ~joined_ts_notna["value_l"], "value_l"
-        ]
-        self.ts_diffs = joined_ts.groupby("variable fips".split()).apply(timeseries_diff)
-
-
-def timeseries_diff(ts: pd.DataFrame) -> float:
-    try:
-        ts = ts.droplevel(["variable", CommonFields.FIPS])
-        right = ts["value_r"]
-        left = ts["value_l"]
-        start = max(right.idxmin(), left.idxmin())
-        end = min(right.idxmax(), left.idxmax())
-        if start <= end:
-            right_common_ts = right.loc[start:end].interpolate(method="time")
-            left_common_ts = left.loc[start:end].interpolate(method="time")
-            diff = (
-                (right_common_ts - left_common_ts).abs() / ((right_common_ts + left_common_ts) / 2)
-            ).mean()
-            if diff > 0.01:
-                print(ts)
-                print(f"from {start} to {end}")
-            return pd.Series(
-                [diff, len(right_common_ts), True], index=["diff", "points_overlap", "has_overlap"]
-            )
-        else:
-            return pd.Series([1.0, 0, False], index=["diff", "points_overlap", "has_overlap"])
-    except:
-        ts.info()
-        print(ts)
-        return float("NaN")
 
 
 @main.command()
