+++ test/common_df_diff_test.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from cli.utils import DatasetDiff
+from libs.qa.common_df_diff import DatasetDiff
 from covidactnow.datapublic.common_fields import COMMON_FIELDS_TIMESERIES_KEYS
